don't clear scons cache in CI
old-commit-hash: cf72a4eb34287218f385e75849902755203d55ec

+++ selfdrive/manager.py
@@ -123,12 +123,16 @@ if not prebuilt:
       compile_output += r
 
       if retry:
-        print("scons build failed, cleaning in")
-        for i in range(3, -1, -1):
-          print("....%d" % i)
-          time.sleep(1)
-        subprocess.check_call(["scons", "-c"], cwd=BASEDIR, env=env)
-        shutil.rmtree("/tmp/scons_cache")
+        if not os.getenv("CI"):
+          print("scons build failed, cleaning in")
+          for i in range(3, -1, -1):
+            print("....%d" % i)
+            time.sleep(1)
+          subprocess.check_call(["scons", "-c"], cwd=BASEDIR, env=env)
+          shutil.rmtree("/tmp/scons_cache")
+        else:
+          print("scons build failed after retry")
+          sys.exit(1)
       else:
         # Build failed log errors
         errors = [line.decode('utf8', 'replace') for line in compile_output
